+++ src/library/adash-messages.ads
@@ -300,6 +300,7 @@ package Adash.Messages is
       Msg_List_Arguments,
       Msg_List_Choices,
       Msg_List_Handlers,
+      Msg_List_Nesting,
       Msg_Case_Not_Discrete,
       Msg_Case_Choice_Not_Static,
       Msg_Case_Choice_Covered_Twice,